requirements.txt
flask
requests
beautifulsoup4
kafka-python
